Services splitted to different .py scripts and replaced in services module
