+++ comm_app/services/filter_service.py
@@ -1,7 +1,3 @@
-import csv
-
-from django.http import HttpResponse
-
 from comm_app.models import Comment
 
 
@@ -27,19 +23,3 @@ def get_queryset_filter(entity_name, object_id, username, start_date, end_date, 
 
     else:
         return Comment.objects.all()
-
-
-def export_to_csv(exporting_queryset):
-    output = []
-    response = HttpResponse(content_type='text/csv')
-    response.write(u'\ufeff'.encode('utf8'))
-    writer = csv.writer(response)
-    writer.writerow(['Author', 'Content', 'Creation Date', 'Related Entity', 'Object ID'])
-    for comment in exporting_queryset:
-        output.append([comment.author,
-                       comment.content,
-                       comment.creation_date,
-                       comment.content_type.model,
-                       comment.object_id])
-    writer.writerows(output)
-    return response
