+++ comm_app/views.py
@@ -5,7 +5,7 @@ from rest_framework.viewsets import GenericViewSet
 
 from comm_app.models import *
 from comm_app.serializers import AllCommentsSerializer
-from comm_app.services import *
+from comm_app.services import export_service, filter_service
 
 
 class CommentsViewSet(mixins.CreateModelMixin,
@@ -20,7 +20,7 @@ class CommentsViewSet(mixins.CreateModelMixin,
         start_date = self.request.query_params.get('start_date')
         end_date = self.request.query_params.get('end_date')
 
-        return get_queryset_filter(entity_name, object_id, username, start_date, end_date, export)
+        return filter_service.get_queryset_filter(entity_name, object_id, username, start_date, end_date, export)
 
     def perform_create(self, serializer):  # Autocomplete of "Parent" field by content_type and id
         content_type_id = self.request.data.get('content_type')
@@ -40,4 +40,4 @@ class CommentsViewSet(mixins.CreateModelMixin,
     @action(detail=False, methods=['get'])
     def csv_export(self, request):
         exporting_queryset = self.get_queryset(export=True)
-        return export_to_csv(exporting_queryset)
+        return export_service.export_to_csv(exporting_queryset)
